+++ src/components/ExitTicketForm/ExitTicketForm.css
@@ -7,7 +7,7 @@
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
   opacity: 0;
-  visibility: hidden;
+  /* visibility: hidden; */
   transform: scale(1.1);
 }
