+++ src/components/LearningTargetForm/LearningTargetForm.js
@@ -1,57 +1,18 @@
 import React from 'react';
 import './LearningTargetForm.css';
-import config from '../../config';
-import TokenService from '../../Services/token-service';
 
 class LearningTargetForm extends React.Component {
   state = {
     error: null,
-    learningTarget: '',
   }
   staticDefaultProps={}
 
-  // Updates state with every user input change
-  handleChange = (e) => {
-    this.setState({
-      learningTarget: e.target.value,
-    })
-  }
-
   handleSubmit = (e) => {
     e.preventDefault();
-    const { learningTarget } = e.target;
-    const data = {
-      goal_title: learningTarget.value,
-      ...this.state.exitTicketData,
-    }
-    fetch(`${config.API_ENDPOINT}/goals/class/${this.props.classId}`, {
-      method: 'POST',
-      body: JSON.stringify(data),
-      headers: {
-        'content-type': 'application/json',
-        authorization: `bearer ${TokenService.getAuthToken()}`
-      }
-    }).then((res) => {
-      if(!res){
-        return res.json().then(e => Promise.reject(e));
-      }
-      return res.json();
-    }).then(res => {
-      this.setState({
-        learningTarget: ''
-      })
-      // const { history } = this.props
-      // history.push('/session')
-    })
-    .catch(error => {
-      console.error({ error })
-    })
+    this.props.showModal();
   }
-
   render() {
-    // console.log(
-    //   'PROPS', this.props
-    // )
+    
     return (
       <div className='learning-target-container'>
       <form className='learning-target-form'
@@ -59,8 +20,8 @@ class LearningTargetForm extends React.Component {
             <label htmlFor='learningTarget'>Learning Target:</label>
             <textarea 
               id='learning-target'
-              onChange={this.handleChange}
-              value={this.state.learningTarget}
+              onChange={this.props.handleChange}
+              value={this.props.learningTarget}
               name='learningTarget'
               aria-label='learning target'
               aria-required='true'
